Add classes for calculator

+++ src/test/java/com/github/jinahya/jsonrpc/bind/v2/examples/jsonrpc_org/v2/NamedParametersRequestTest.java
@@ -74,7 +74,7 @@ class NamedParametersRequestTest
                     assertEquals("subtract", v.getMethod());
                     assertEquals(23, v.getParams().getSubtrahend());
                     assertEquals(42, v.getParams().getMinuend());
-                    assertEquals(3, v.getId());
+                    assertEquals(Integer.valueOf(3), v.getId());
                 }
         );
         JacksonUtils.readTreeFromResource(
@@ -100,7 +100,7 @@ class NamedParametersRequestTest
                     assertEquals("subtract", v.getMethod());
                     assertEquals(23, v.getParams().getSubtrahend());
                     assertEquals(42, v.getParams().getMinuend());
-                    assertEquals(4, v.getId());
+                    assertEquals(4, (int) v.getId());
                 }
         );
     }
