+++ src/test/java/com/github/jinahya/jsonrpc/bind/v2/examples/jsonrpc_org/v2/NamedParametersResponseTest.java
@@ -41,16 +41,16 @@ class NamedParametersResponseTest
     @Test
     void named_parameters_01_response() throws IOException {
         acceptValueFromResource("/examples/jsonrpc_org/v2/named_parameters_01_response.json", v -> {
-            assertEquals(19, v.getResult());
-            assertEquals(3, v.getId());
+            assertEquals(Integer.valueOf(19), v.getResult());
+            assertEquals(Integer.valueOf(3), v.getId());
         });
     }
 
     @Test
     void named_parameters_02_response() throws IOException {
         acceptValueFromResource("/examples/jsonrpc_org/v2/named_parameters_02_response.json", v -> {
-            assertEquals(19, v.getResult());
-            assertEquals(4, v.getId());
+            assertEquals(19, (int) v.getResult());
+            assertEquals(4, (int) v.getId());
         });
     }
 }
