+++ src/test/java/com/github/jinahya/jsonrpc/bind/v2/examples/jsonrpc_org/v2/PositionalParametersRequestTest.java
@@ -42,11 +42,13 @@ class PositionalParametersRequestTest extends RequestObjectTest<PositionalParame
 
     @Test
     void positional_parameters_01_request() throws IOException {
-        acceptValueFromResource("/examples/jsonrpc_org/v2/positional_parameters_01_request.json", v -> {
-            assertEquals("subtract", v.getMethod());
-            assertIterableEquals(asList(42, 23), v.getParams());
-            assertEquals(1, v.getId());
-        });
+        acceptValueFromResource(
+                "/examples/jsonrpc_org/v2/positional_parameters_01_request.json",
+                v -> {
+                    assertEquals("subtract", v.getMethod());
+                    assertIterableEquals(asList(42, 23), v.getParams());
+                    assertEquals(1, (int) v.getId());
+                });
     }
 
     @Test
@@ -54,7 +56,7 @@ class PositionalParametersRequestTest extends RequestObjectTest<PositionalParame
         acceptValueFromResource("/examples/jsonrpc_org/v2/positional_parameters_02_request.json", v -> {
             assertEquals("subtract", v.getMethod());
             assertIterableEquals(asList(23, 42), v.getParams());
-            assertEquals(2, v.getId());
+            assertEquals(2, (int) v.getId());
         });
     }
 }
